enhance(explorer): show feedback for unsupported explorers
